# gbt/openapi.yaml
openapi: 3.0.3
info:
  title: Records_AI UPAP API
  description: |
    Official API for the Records_AI platform.
    All operations strictly follow the UPAP pipeline:
    Upload → Archive → Process → Publish.

    No automatic progression is allowed.
    Each stage must be triggered explicitly.
  version: "1.0.0"
servers:
  - url: https://api.records-ai.com
    description: Production server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
    RecordResponse:
      type: object
      properties:
        record_id:
          type: string
          format: uuid
        stage:
          type: string
          enum: [uploaded, archived, processed, published]
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
paths:
  /upap/upload:
    post:
      summary: Upload a file to Records_AI
      description: |
        First step of the UPAP pipeline.
        Accepts a file and optional metadata.
      operationId: uploadRecord
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                metadata:
                  type: string
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecordResponse"
        "400":
          description: Invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...